renamed and removed deprecated file
